- Update the "GetThreadMessagesAsync" function of the "MessageRepository" so that it does not make a big query for the messages and updates on the fly the "DateRead" field of the the database message - Update the "MessageRepository" by adding the new "GetGroupConnection" that returns the group to which the requested connection belongs to - Update the "PresenceHub" so that on a new client connection it sends the "UserIsOnline" notification only when for the first time from one device a member is connected (not if he was already connected from another device) and send the "GetOnlineUsers" notification only to the "Caller" to update his list on the client app - Update the "PresenceHub" so that on a client disconnection it sends the "UserIsOffliine" notification to "Others" and not to "All" only when all devices of that member are disconnected (not connected from any device) - Update the "MessageHub" on connection function to send the "UpdateGroup" notification to the whole group when a new member is connected and the "ReceivedMessageThread" only to the "Caller" to receive all messages from that thread - Update the "MessageHub" on discconnection function to send the "UpdateGroup" notification to the whole group when a new member is disconnected - Update on the Client app respective the "presence" service so that it updates the users online list when the "UserIsOnline" or "UserIsOffline" events are catched - Update on the Client app respective the "message" service so that it updates the internal list with all messages from the thread to "read" when the other member of the group is connected (joined the MessageHub by opening his "Messages" tab)

+++ DatingApp/DatingApp.API/SignalR/MessageHub.cs
@@ -19,15 +19,18 @@ IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
             throw new HubException("Cannot join group");
         var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToMessageGroupAsync(groupName);
+        var group = await AddToMessageGroupAsync(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
         var messages = await messageRepository.GetThreadMessagesAsync(Context.User.GetUsername(), otherUser!);
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -86,7 +89,7 @@ IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
         }
     }
 
-    private async Task<bool> AddToMessageGroupAsync(string groupName)
+    private async Task<Entities.Group> AddToMessageGroupAsync(string groupName)
     {
         var username = Context.User?.GetUsername() ?? throw new HubException("Cannot get current username");
 
@@ -102,17 +105,20 @@ IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
         var connection = new Entities.Connection { ConnectionId = Context.ConnectionId, Username = username };
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync()) return group;            
+        throw new HubException("Failed to join group");
     }
 
-    private async Task RemoveFromMessageGroup()
+    private async Task<Entities.Group> RemoveFromMessageGroup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if (connection != null)
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(c => c.ConnectionId == Context.ConnectionId);
+        if (connection != null && group != null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
         }
+        throw new HubException("Failed to remove from group");
     }
 
     private string GetGroupName(string? caller, string? other)
